Reloading partially done MD in Tully

+++ nff/md/nvt.py
@@ -28,9 +28,10 @@ class NoseHoover(MolecularDynamics):
                  max_steps=None,
                  nbr_update_period=20,
                  append_trajectory=True,
+                 remove_old_file=True,
                  **kwargs):
 
-        if os.path.isfile(str(trajectory)):
+        if os.path.isfile(str(trajectory)) and remove_old_file:
             os.remove(trajectory)
 
         MolecularDynamics.__init__(self,
